refactor: move props to separate files
