+++ src/components/BlogScreenshotImage/types/IProps.ts
@@ -1,0 +1,10 @@
+// types
+import type { IThemedImageSources } from '@site/src/types';
+
+interface IProps {
+  label: string;
+  maxWidth?: string;
+  imageSources: IThemedImageSources;
+}
+
+export default IProps;
