+++ src/components/BlogScreenshotImage/types/index.ts
@@ -1,0 +1,1 @@
+export { default as IProps } from './IProps';
