+++ src/components/Button/Button.tsx
@@ -1,14 +1,16 @@
 import clsx from 'clsx';
-import React, { FC, ButtonHTMLAttributes } from 'react';
+import React, { type FC, type ButtonHTMLAttributes } from 'react';
 
 // styles
 import styles from './styles.module.scss';
 
-interface IProps extends ButtonHTMLAttributes<HTMLButtonElement> {
-  fullWidth?: boolean;
-}
+// types
+import type { IProps } from './types';
 
-const Button: FC<IProps> = ({ fullWidth = false, ...buttonProps }: IProps) => {
+const Button: FC<IProps & ButtonHTMLAttributes<HTMLButtonElement>> = ({
+  fullWidth = false,
+  ...buttonProps
+}) => {
   const classNames: string[] = [styles.button];
 
   if (fullWidth) {
