+++ src/components/Button/types/IProps.ts
@@ -1,0 +1,5 @@
+interface IProps {
+  fullWidth?: boolean;
+}
+
+export default IProps;
