+++ src/components/Carousel/types/IProps.ts
@@ -1,0 +1,8 @@
+// types
+import type ICarouselItemProps from './ICarouselItemProps';
+
+interface IProps {
+  items: ICarouselItemProps[];
+}
+
+export default IProps;
