+++ src/components/Carousel/types/index.ts
@@ -1,1 +1,2 @@
 export { default as ICarouselItemProps } from './ICarouselItemProps';
+export { default as IProps } from './IProps';
