+++ src/components/CenteredImage/types/IProps.ts
@@ -1,0 +1,7 @@
+interface IProps {
+  label: string;
+  maxWidth?: string;
+  src: string;
+}
+
+export default IProps;
