+++ src/components/Heading/types/IProps.ts
@@ -1,0 +1,5 @@
+interface IProps {
+  level?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
+}
+
+export default IProps;
