+++ src/components/LinkButton/types/IProps.ts
@@ -1,0 +1,6 @@
+interface IProps {
+  disabled?: boolean;
+  fullWidth?: boolean;
+}
+
+export default IProps;
