+++ src/components/Section/types/IProps.ts
@@ -1,0 +1,7 @@
+interface IProps {
+  id: string;
+  title: string;
+  variant?: 'accent' | 'primary';
+}
+
+export default IProps;
