+++ src/components/Text/types/IProps.ts
@@ -1,0 +1,5 @@
+interface IProps {
+  bold?: boolean;
+}
+
+export default IProps;
